+ Corection de l'affichage des erreurs de valeur des programmes     + Ajout d'un champ permettant de définir le début et la fin d'un cycle     + Ajout de la vérification à la saisie des valeurs pour les heures et les données rentrées dans le formulaire des programmes     + Ajout de listes pour choisir la prise impactée par l'import ou l'export des programmes     + Correction du lien du bouton permettant de passer d'un programme d'une prise à la configuration de la prise

+++ trunk/01_software/02_src/joomla/main/libs/js/cultibox.js
@@ -1338,7 +1338,7 @@ $(document).ready(function() {
                         data: {value:$("#value_program").val(),type:'value_program',plug_type:plugs_infoJS[$('#selected_plug option:selected').val()-1]['PLUG_TYPE']}
                     }).done(function (data) {
                         if(data!=1) {
-                            $("#error_value_program").html("<img src='/cultibox/main/libs/img/arrow_error.png' alt=''>"+error_valueJS[data]);
+                            $("#error_value_program").html("<img src='/cultibox/main/libs/img/arrow_error.png' alt=''>"+error_valueJS[data.toInt()]);
                             $("#error_value_program").show(700);
                             checked=false;
                         } 
@@ -1561,7 +1561,7 @@ $(document).ready(function() {
                     data: {value:$("#value_program").val(),type:'value_program',plug_type:plug_type}
                 }).done(function (data) {
                     if(data!=1) {
-                        $("#error_value_program").html("<img src='/cultibox/main/libs/img/arrow_error.png' alt=''>"+error_valueJS[data]);
+                        $("#error_value_program").html("<img src='/cultibox/main/libs/img/arrow_error.png' alt=''>"+error_valueJS[data.toInt()]);
                         $("#error_value_program").show(700);
                         checked=false;
                     }
